% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblematrix-data.R, R/bubblematrix.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{An example data for bubble_matrix()}
\format{
An object of class \code{data.frame} with 3 rows and 5 columns.
}
\usage{
example_data
}
\arguments{
\item{x}{x is a number}
}
\value{
a dataframe
}
\description{
These data were generated from the Manson_VITAL_protocol_NEJM_2019
study.

today,I create my first function,a very usrful function.
}
\details{
This is some description of this function.

you can use this function to caculate x+1,then return the value of x+1.
}
\keyword{datasets}
