% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblematrix.R
\name{bubble_matrix}
\alias{bubble_matrix}
\title{simple function}
\usage{
bubble_matrix(
  data,
  percentile = c(0, 0.25, 0.5, 0.75, 1),
  labels = c("lowest rate", "below rate", "above rate", "highest rate"),
  color = c("#F9DBD3", "#F1B255", "#519F46", "#41B0C3"),
  ploty_ = TRUE,
  title = "",
  subtitle = "",
  caption = ""
)
}
\arguments{
\item{x}{x is a number}
}
\value{
a dataframe
}
\description{
today,I create my first function,a very usrful function.
}
\details{
This is some description of this function.

you can use this function to caculate x+1,then return the value of x+1.
}
