% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblematrix.R
\name{bubble_matrix}
\alias{bubble_matrix}
\title{bubblematrix function}
\usage{
bubble_matrix(
  data,
  percentile = c(0, 0.25, 0.5, 0.75, 1),
  labels = c("lowest rate", "below rate", "above rate", "highest rate"),
  color = c("#F9DBD3", "#F1B255", "#519F46", "#41B0C3"),
  ploty_ = TRUE,
  title = "",
  subtitle = "",
  caption = ""
)
}
\arguments{
\item{data}{the data plot}

\item{percentile}{Determine the sub-locations for the drawing}

\item{labels}{labels of each interval}
}
\value{
None Show the plot
}
\description{
Create your own dynamic or state bubblematrix plot
}
\details{
This is some description of this function.

you can use this function to plot bubblematrix(ploty or not).
}
