% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblematrix-data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{An example data for bubble_matrix()}
\format{
An object of class \code{data.frame} with 3 rows and 5 columns.
}
\usage{
example_data
}
\description{
These data were generated from the Manson_VITAL_protocol_NEJM_2019
study.
 @name example_data
}
\keyword{datasets}
